chore: readme

+++ entity/src/otp.rs
@@ -9,7 +9,7 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: String,
-    pub code: i32,
+    pub code: i16,
     pub email: String,
     pub valid_till: DateTimeWithTimeZone,
     pub purpose: OtpPurpose,
